requirements.txt
requests>=2.31.0
pandas>=2.0.0
openpyxl>=3.1.0
pyyaml>=6.0.0
XlsxWriter>=3.1.0
